Уменьшено количество приведений типов.

+++ Source/DX11VideoProcessor.cpp
@@ -2261,8 +2261,8 @@ HRESULT CDX11VideoProcessor::CopySample(IMediaSample* pSample)
 		}
 
 		BYTE* data = nullptr;
-		const long size = pSample->GetActualDataLength();
-		if (size >= static_cast<long>(abs(m_srcPitch) * m_srcLines) && S_OK == pSample->GetPointer(&data)) {
+		const int size = pSample->GetActualDataLength();
+		if (size >= abs(m_srcPitch) * (int)m_srcLines && S_OK == pSample->GetPointer(&data)) {
 			// do not use UpdateSubresource for D3D11 VP here
 			// because it can cause green screens and freezes on some configurations
 			hr = MemCopyToTexSrcVideo(data, m_srcPitch);
