+++ Source/DX9VideoProcessor.cpp
@@ -1468,8 +1468,8 @@ HRESULT CDX9VideoProcessor::CopySample(IMediaSample* pSample)
 		}
 
 		BYTE* data = nullptr;
-		const long size = pSample->GetActualDataLength();
-		if (size >= static_cast<long>(abs(m_srcPitch) * m_srcLines) && S_OK == pSample->GetPointer(&data)) {
+		const int size = pSample->GetActualDataLength();
+		if (size >= abs(m_srcPitch) * (int)m_srcLines && S_OK == pSample->GetPointer(&data)) {
 			if (m_srcParams.cformat == CF_NONE) {
 				return E_FAIL;
 			}
